route/requirements.txt
Flask==3.1.0
Flask-RESTful==0.3.10
flask-cors==5.0.1
Werkzeug==3.1.3

pytest==8.3.5
coverage==7.7.1
requests-mock==1.11.0